Extracted user login form to separate box component
