+++ client/components/UserLoginBox/Wrapper.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+
+const Wrapper = styled.div`
+  width: 500px;
+  max-width: 100%;
+  padding: 20px;
+  background-color: ${props => props.theme.color.grey5}
+`
+
+export default Wrapper
